[Global] Favicon selection enhancement (#3233)
Show favicon file name and file title in the favicon selection modal.
Also refactored the component.


[Launchpad---Zesty-io---zesty-pw---Manager.webm](https://github.com/user-attachments/assets/ef3910cc-7a75-471d-b962-61509170d588)
