+++ src/apps/content-editor/src/app/components/FieldTypeMedia.tsx
@@ -967,3 +967,5 @@ export const MediaItem = ({
     </>
   );
 };
+
+FieldTypeMedia.displayName = "FieldTypeMedia";
