+++ src/shell/components/Favicon/favicon.less
@@ -9,7 +9,8 @@
 }
 
 .Modal {
-  max-width: 800px;
+  width: 800px;
+  max-width: none;
   overflow-y: auto;
   max-height: 80vh;
