--- src/shell/components/favicon/index.js
@@ -1,2 +1,0 @@
-import Favicon from "./favicon";
-export default Favicon;
